api/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==0.8.11
pydantic==2.5.0
uvicorn==0.24.0